Fix cancel action when modal closes from overlay
Prior to this change, the modal would close but the confirmer would
never resolve.

This change will cancel the confirmer when the modal closes. There
is no action other then close which will be triggered when the user
closes the modal for any reason including an overlay click or escape
etc.

Since a confirmer can only have on resolution (first come first serve)
repeated resolves will be ignored.